Add in DAGMan class to handle DAG functionality.
This allows for implementation of DAGs, using the existing
Job and JobSet classes. Some modifications to JobSet are required,
to avoid adding all jobs to the submit file (or writing out lots of
identical submit files), but this is transparent to the user.

Note that checking that the jobs are acyclic isn't implemented yet!